Refactoring 'validate.js' from Registros
