+++ js/validateLogin/events.js
@@ -1,6 +1,5 @@
 //Importaciones
 import validateLogin from "./validate.js";
-import dataManagement from "../dataManagement/dataManagement.js";
 
 //Llamando al button
 const btnIngresar = document.querySelector('[Ingresar]');
